refactoring memory rep
